Avoid polute Python root logger because of importing pytorch module
We used to configure root logger in RPC module. A stream handler is added to `root.handlers`. This is not desired behavior for pytorch users. We should instead keep the root logger handler list untouched.

We can configure the logger local to the rpc module, set it's log level, so it doesn't use it's ancestor, which is usually the root which has no stream handlers in most cases.
https://docs.python.org/3/library/logging.html#logging.Logger.setLevel

And add a stream handler to make it output to stdout, even if the root handlers is not configured and has an empty list.
https://docs.python.org/3/library/logging.html#logging.Logger.addHandler
https://docs.python.org/3/library/logging.handlers.html#logging.StreamHandler

Differential Revision: [D7677493](https://our.internmc.facebook.com/intern/diff/D7677493/)

ghstack-source-id: 100291775
Pull Request resolved: https://github.com/pytorch/pytorch/pull/34871